pyproject.toml
[project]
name = "scada-chatbot"
version = "0.1.0"
description = "RAG Chatbot using Streamlit, LlamaIndex, PostgreSQL, and Ollama"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.28.0",
    "llama-index>=0.9.0",
    "llama-index-embeddings-ollama>=0.1.0",
    "llama-index-llms-ollama>=0.1.0",
    "llama-index-vector-stores-postgres>=0.1.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    "pgvector>=0.2.0",
    "pypdf>=3.17.0",
    "python-dotenv>=1.0.0",
    "ollama>=0.1.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "loguru>=0.7.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "requests>=2.32.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
